refactors app/ and fixes most eslint errors
