--- app/routes/scrapes.server.route.js
@@ -1,7 +1,0 @@
-'use strict';
-var Scrapes = require('../controllers/scrapes.server.controller');
-
-module.exports = function(app) {
-  app.route('/scrapes')
-    .get(Scrapes.list);
-};
