--- app/routes/subscribers.server.route.js
@@ -1,7 +1,0 @@
-'use strict';
-var Subscribers = require('../controllers/subscribers.server.controller');
-
-module.exports = function(app) {
-  app.route('/subscribers')
-    .post(Subscribers.update);
-};
