+++ package.json
@@ -19,13 +19,16 @@
   },
   "homepage": "https://github.com/casche/gitajob",
   "dependencies": {
+    "async": "^1.4.2",
     "body-parser": "^1.12.2",
     "bower": "^1.4.1",
     "cheerio": "latest",
+    "colors": "^1.1.2",
     "cron": "^1.0.9",
     "ejs": "^2.3.1",
     "express": "^4.12.3",
     "glob": "^5.0.5",
+    "http": "0.0.0",
     "mailgun-js": "^0.6.9",
     "method-override": "^2.3.2",
     "mongoose": "^4.0.1",
