+++ public/js/app.js
@@ -1,1 +1,2 @@
+/* global angular */
 angular.module('app', ['ngRoute', 'appRoutes', 'ngResource', 'ui.bootstrap']);
