+++ public/js/controllers/footer.controller.js
@@ -1,1 +1,3 @@
+/* global angular */
+
 angular.module('app').controller('FooterController', ['$scope', '$location', function($scope, $location) {}]);
