+++ public/js/services/subscribers.client.service.js
@@ -1,3 +1,4 @@
+/* global angular */
 'use strict';
 
 angular.module('app').factory('Subscribers', ['$resource',
